+++ packages/terra-table/src/subcomponents/Section.jsx
@@ -5,6 +5,7 @@ import PropTypes from 'prop-types';
 import SectionHeader from 'terra-section-header';
 
 import ThemeContext from 'terra-theme-context';
+import GridContext, { GridConstants } from '../utils/GridContext';
 
 import Row from './Row';
 import rowShape from '../proptypes/rowShape';
@@ -122,6 +123,9 @@ function Section(props) {
 
   const theme = useContext(ThemeContext);
 
+  const gridContext = useContext(GridContext);
+  const isGridContext = gridContext.role === GridConstants.GRID;
+
   const handleMouseDown = (event) => {
     onSectionSelect(id);
     event.stopPropagation();
@@ -134,6 +138,7 @@ function Section(props) {
       <tr
         aria-rowindex={sectionRowIndex}
         className={cx('header')}
+        data-section-id={id}
       >
         <th
           id={`${tableId}-${id}`}
@@ -141,10 +146,12 @@ function Section(props) {
           colSpan={displayedColumns.length}
           role="columnheader"
           scope="col"
+          tabIndex={isGridContext && !(isCollapsible && onSectionSelect) ? -1 : undefined}
         >
           <SectionHeader
             text={text}
             isOpen={!isCollapsed}
+            isTitleFixed
             onClick={isCollapsible && onSectionSelect ? handleMouseDown : undefined}
           />
         </th>